+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/JavaNetHttpPollingClientFactory.java
@@ -5,32 +5,29 @@ import java.net.http.HttpClient;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
 import org.apache.flink.table.data.RowData;
 
-import com.getindata.connectors.http.LookupQueryCreator;
-import com.getindata.connectors.http.internal.HeaderPreprocessor;
 import com.getindata.connectors.http.internal.PollingClientFactory;
-import com.getindata.connectors.http.internal.utils.HttpHeaderUtils;
 import com.getindata.connectors.http.internal.utils.JavaNetHttpClientFactory;
 
 public class JavaNetHttpPollingClientFactory implements PollingClientFactory<RowData> {
 
+    private final HttpRequestFactory requestFactory;
+
+    public JavaNetHttpPollingClientFactory(HttpRequestFactory requestFactory) {
+        this.requestFactory = requestFactory;
+    }
+
     @Override
     public JavaNetHttpPollingClient createPollClient(
-        HttpLookupConfig options,
-        DeserializationSchema<RowData> schemaDecoder,
-        LookupQueryCreator lookupQueryCreator) {
+            HttpLookupConfig options,
+            DeserializationSchema<RowData> schemaDecoder) {
 
         HttpClient httpClient = JavaNetHttpClientFactory.createClient(options.getProperties());
-
-        // TODO Consider this to be injected as method argument or factory field
-        //  so user could set this using API.
-        HeaderPreprocessor headerPreprocessor = HttpHeaderUtils.createDefaultHeaderPreprocessor();
 
         return new JavaNetHttpPollingClient(
             httpClient,
             schemaDecoder,
             options,
-            lookupQueryCreator,
-            headerPreprocessor
+            requestFactory
         );
     }
 }