+++ src/test/java/com/getindata/connectors/http/internal/table/lookup/JavaNetHttpPollingClientConnectionTest.java
@@ -7,6 +7,7 @@ import java.util.Properties;
 import java.util.stream.Stream;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
+import com.github.tomakehurst.wiremock.client.MappingBuilder;
 import com.github.tomakehurst.wiremock.matching.RequestPatternBuilder;
 import com.github.tomakehurst.wiremock.stubbing.StubMapping;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
@@ -32,13 +33,21 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
+import static com.github.tomakehurst.wiremock.client.WireMock.equalToJson;
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
+import static com.github.tomakehurst.wiremock.client.WireMock.post;
+import static com.github.tomakehurst.wiremock.client.WireMock.postRequestedFor;
+import static com.github.tomakehurst.wiremock.client.WireMock.put;
+import static com.github.tomakehurst.wiremock.client.WireMock.putRequestedFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.fail;
 
 import com.getindata.connectors.http.LookupArg;
 import com.getindata.connectors.http.internal.config.HttpConnectorConfigConstants;
 import com.getindata.connectors.http.internal.table.lookup.querycreators.GenericGetQueryCreator;
+import com.getindata.connectors.http.internal.table.lookup.querycreators.GenericJsonQueryCreator;
+import com.getindata.connectors.http.internal.utils.HttpHeaderUtils;
 import static com.getindata.connectors.http.TestHelper.readTestFile;
 
 @ExtendWith(MockitoExtension.class)
@@ -47,6 +56,8 @@ class JavaNetHttpPollingClientConnectionTest {
     private static final String SAMPLES_FOLDER = "/http/";
 
     private static final List<String> LOOKUP_KEYS = List.of("id", "uuid");
+
+    private static final String ENDPOINT = "/service";
 
     private static WireMockServer wireMockServer;
 
@@ -57,7 +68,7 @@ class JavaNetHttpPollingClientConnectionTest {
 
     private StubMapping stubMapping;
 
-    private JavaNetHttpPollingClientFactory pollingClientFactory;
+    private Properties properties;
 
     @BeforeAll
     public static void setUpAll() {
@@ -76,7 +87,12 @@ class JavaNetHttpPollingClientConnectionTest {
     public void setUp() {
         int[][] lookupKey = {{}};
         this.dynamicTableSourceContext = new LookupRuntimeProviderContext(lookupKey);
-        this.pollingClientFactory = new JavaNetHttpPollingClientFactory();
+
+        this.properties = new Properties();
+        this.properties.setProperty(
+            HttpConnectorConfigConstants.LOOKUP_SOURCE_HEADER_PREFIX + "Content-Type",
+            "application/json"
+        );
     }
 
     @AfterEach
@@ -88,10 +104,12 @@ class JavaNetHttpPollingClientConnectionTest {
 
     @Test
     void shouldQuery200WithParams() {
-        stubMapping = setupServerStub(200);
 
+        // GIVEN
+        this.stubMapping = setUpServerStub(200);
         JavaNetHttpPollingClient pollingClient = setUpPollingClient(getBaseUrl(), LOOKUP_KEYS);
 
+        // WHEN
         RowData result = pollingClient.pull(
             List.of(
                 new LookupArg("id", "1"),
@@ -99,7 +117,51 @@ class JavaNetHttpPollingClientConnectionTest {
             )
         ).orElseThrow();
 
+        // THEN
         wireMockServer.verify(RequestPatternBuilder.forCustomMatcher(stubMapping.getRequest()));
+
+        assertThat(result.getArity()).isEqualTo(4);
+        assertThat(result.getString(1)
+            .toString()).isEqualTo("Returned HTTP message for parameter PARAM, COUNTER");
+
+        RowData detailsRow = result.getRow(3, 2);
+        assertThat(detailsRow.getBoolean(0)).isEqualTo(true);
+
+        RowData nestedDetailsRow = detailsRow.getRow(1, 1);
+        assertThat(nestedDetailsRow.getString(0).toString()).isEqualTo("$1,729.34");
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"PUT", "POST"})
+    void shouldQuery200WithBodyParams(String methodName) {
+
+        // GIVEN
+        this.stubMapping = setUpServerBodyStub(methodName);
+        JavaNetHttpPollingClient pollingClient =
+            setUpPollingClient(
+                getBaseUrl(),
+                LOOKUP_KEYS,
+                properties,
+                setUpBodyRequestFactory(methodName, properties)
+            );
+
+        // WHEN
+        RowData result = pollingClient.pull(
+            List.of(
+                new LookupArg("id", "1"),
+                new LookupArg("uuid", "2")
+            )
+        ).orElseThrow();
+
+        // THEN
+        wireMockServer.verify(RequestPatternBuilder.forCustomMatcher(stubMapping.getRequest()));
+        if (methodName.equalsIgnoreCase("POST")) {
+            wireMockServer.verify(postRequestedFor(urlEqualTo(ENDPOINT)));
+        } else if (methodName.equalsIgnoreCase("PUT")) {
+            wireMockServer.verify(putRequestedFor(urlEqualTo(ENDPOINT)));
+        } else {
+            fail("Unexpected REST method.");
+        }
 
         assertThat(result.getArity()).isEqualTo(4);
         assertThat(result.getString(1)
@@ -125,14 +187,17 @@ class JavaNetHttpPollingClientConnectionTest {
     void shouldHandleCodeBasedOnConfiguration(
             Properties properties,
             boolean isExpectedResponseEmpty) {
-        stubMapping = setupServerStub(201);
 
+        // GIVEN
+        this.stubMapping = setUpServerStub(201);
         JavaNetHttpPollingClient pollingClient = setUpPollingClient(
             getBaseUrl(),
             LOOKUP_KEYS,
-            properties
+            properties,
+            setUpGetRequestFactory(properties)
         );
 
+        // WHEN
         Optional<RowData> poll = pollingClient.pull(
             List.of(
                 new LookupArg("id", "1"),
@@ -140,15 +205,18 @@ class JavaNetHttpPollingClientConnectionTest {
             )
         );
 
+        // THEN
         assertThat(poll.isEmpty()).isEqualTo(isExpectedResponseEmpty);
     }
 
     @Test
     void shouldHandleServerError() {
-        stubMapping = setupServerStub(500);
 
+        // GIVEN
+        this.stubMapping = setUpServerStub(500);
         JavaNetHttpPollingClient pollingClient = setUpPollingClient(getBaseUrl(), LOOKUP_KEYS);
 
+        // WHEN
         Optional<RowData> poll = pollingClient.pull(
             List.of(
                 new LookupArg("id", "1"),
@@ -156,6 +224,7 @@ class JavaNetHttpPollingClientConnectionTest {
             )
         );
 
+        // THEN
         wireMockServer.verify(RequestPatternBuilder.forCustomMatcher(stubMapping.getRequest()));
 
         assertThat(poll.isEmpty()).isTrue();
@@ -163,13 +232,16 @@ class JavaNetHttpPollingClientConnectionTest {
 
     @Test
     void shouldProcessWithMissingArguments() {
-        stubMapping = setupServerStub(200);
 
+        // GIVEN
+        this.stubMapping = setUpServerStub(200);
         JavaNetHttpPollingClient pollingClient =
             setUpPollingClient(getBaseUrl(), LOOKUP_KEYS);
 
+        // WHEN
         Optional<RowData> poll = pollingClient.pull(Collections.emptyList());
 
+        // THEN
         assertThat(poll.isEmpty()).isTrue();
     }
 
@@ -180,10 +252,10 @@ class JavaNetHttpPollingClientConnectionTest {
     })
     public void shouldConnectWithBasicAuth(String authorizationHeaderValue) {
 
-        stubMapping = setupServerStubForBasicAuth();
+        // GIVEN
+        this.stubMapping = setupServerStubForBasicAuth();
 
         Properties properties = new Properties();
-
         properties.setProperty(
             HttpConnectorConfigConstants.LOOKUP_SOURCE_HEADER_PREFIX + "Content-Type",
             "application/json"
@@ -197,9 +269,11 @@ class JavaNetHttpPollingClientConnectionTest {
         JavaNetHttpPollingClient pollingClient = setUpPollingClient(
             getBaseUrl(),
             LOOKUP_KEYS,
-            properties
+            properties,
+            setUpGetRequestFactory(properties)
         );
 
+        // WHEN
         RowData result = pollingClient.pull(
             List.of(
                 new LookupArg("id", "1"),
@@ -207,6 +281,7 @@ class JavaNetHttpPollingClientConnectionTest {
             )
         ).orElseThrow();
 
+        // THEN
         wireMockServer.verify(RequestPatternBuilder.forCustomMatcher(stubMapping.getRequest()));
 
         assertThat(result.getArity()).isEqualTo(4);
@@ -221,7 +296,7 @@ class JavaNetHttpPollingClientConnectionTest {
     }
 
     private String getBaseUrl() {
-        return wireMockServer.baseUrl() + "/service";
+        return wireMockServer.baseUrl() + ENDPOINT;
     }
 
     public JavaNetHttpPollingClient setUpPollingClient(String url, List<String> arguments) {
@@ -232,13 +307,40 @@ class JavaNetHttpPollingClientConnectionTest {
             "application/json"
         );
 
-        return setUpPollingClient(url, arguments, properties);
+        return setUpPollingClient(url, arguments, properties, setUpGetRequestFactory(properties));
+    }
+
+    private GetRequestFactory setUpGetRequestFactory(Properties properties) {
+        return new GetRequestFactory(
+            new GenericGetQueryCreator(),
+            HttpHeaderUtils.createDefaultHeaderPreprocessor(),
+            HttpLookupConfig.builder()
+                .url(getBaseUrl())
+                .properties(properties)
+                .build()
+        );
+    }
+
+    private BodyBasedRequestFactory setUpBodyRequestFactory(
+            String methodName,
+            Properties properties) {
+        return new BodyBasedRequestFactory(
+            methodName,
+            new GenericJsonQueryCreator(),
+            HttpHeaderUtils.createDefaultHeaderPreprocessor(),
+            HttpLookupConfig.builder()
+                .url(getBaseUrl())
+                .properties(properties)
+                .build()
+        );
     }
 
     private JavaNetHttpPollingClient setUpPollingClient(
             String url,
             List<String> arguments,
-            Properties properties) {
+            Properties properties,
+            HttpRequestFactory requestFactory) {
+
         HttpLookupConfig lookupConfig = HttpLookupConfig.builder()
             .url(url)
             .arguments(arguments)
@@ -262,13 +364,14 @@ class JavaNetHttpPollingClientConnectionTest {
                 .createDecodingFormat(dynamicTableFactoryContext, new Configuration())
                 .createRuntimeDecoder(dynamicTableSourceContext, physicalDataType);
 
-        return pollingClientFactory
-            .createPollClient(lookupConfig, schemaDecoder, new GenericGetQueryCreator());
+        JavaNetHttpPollingClientFactory pollingClientFactory =
+            new JavaNetHttpPollingClientFactory(requestFactory);
+        return pollingClientFactory.createPollClient(lookupConfig, schemaDecoder);
     }
 
-    private StubMapping setupServerStub(int status) {
+    private StubMapping setUpServerStub(int status) {
         return wireMockServer.stubFor(
-            get(urlEqualTo("/service?id=1&uuid=2"))
+            get(urlEqualTo(ENDPOINT + "?id=1&uuid=2"))
                 .withHeader("Content-Type", equalTo("application/json"))
                 .willReturn(
                     aResponse()
@@ -276,8 +379,23 @@ class JavaNetHttpPollingClientConnectionTest {
                         .withBody(readTestFile(SAMPLES_FOLDER + "HttpResult.json"))));
     }
 
+    private StubMapping setUpServerBodyStub(String methodName) {
+        MappingBuilder methodStub = (methodName.equalsIgnoreCase("PUT") ?
+            put(urlEqualTo(ENDPOINT)) :
+            post(urlEqualTo(ENDPOINT)));
+
+        return wireMockServer.stubFor(
+            methodStub
+                .withHeader("Content-Type", equalTo("application/json"))
+                .withRequestBody(equalToJson("{\"id\" : \"1\", \"uuid\" : \"2\"}"))
+                .willReturn(
+                    aResponse()
+                        .withStatus(200)
+                        .withBody(readTestFile(SAMPLES_FOLDER + "HttpResult.json"))));
+    }
+
     private StubMapping setupServerStubForBasicAuth() {
-        return wireMockServer.stubFor(get(urlEqualTo("/service?id=1&uuid=2"))
+        return wireMockServer.stubFor(get(urlEqualTo(ENDPOINT + "?id=1&uuid=2"))
                 .withHeader("Content-Type", equalTo("application/json"))
                 .withBasicAuth("user", "password")
                 .willReturn(